some server side clean up

+++ server/routes/surgery.ts
@@ -2,7 +2,6 @@ import express from 'express';
 import Surgery from '../models/Surgery';
 import authenticateToken from '../middleware/authMiddleware';
 import Patient from '../models/Patient';
-import mongoose from 'mongoose';
 
 const router = express.Router();
 router.use(authenticateToken);
@@ -11,10 +10,13 @@ router.use(authenticateToken);
 router.get('/:patientId', async (req, res) => {
   const { patientId } = req.params;
 
+  if (!patientId) {
+    return res.status(400).json({ message: 'Patient ID is required' });
+  }
+
   try {
     const patient = await Patient.findById(patientId).select('name birthdate age');
     const surgeries = await Surgery.find({ patientId }).select('surgeryDate type surgeon');
-    console.log(`Fetching surgeries for patient ID: ${surgeries}`);
 
     if (!patient) {
       return res.status(404).json({ message: 'Patient not found' });
@@ -42,6 +44,7 @@ router.post('/', async (req, res) => {
     });
 
     const newSurgery = await surgery.save();
+
     res.status(201).json(newSurgery);
   } catch (err) {
     res.status(400).json({ message: err.message });
@@ -51,12 +54,18 @@ router.post('/', async (req, res) => {
 // DELETE (cancel) a surgery by ID
 router.delete('/:surgeryId', async (req, res) => {
   const { surgeryId } = req.params;
+
+  if (!surgeryId) {
+    return res.status(400).json({ message: 'Surgery ID is required' });
+  }
+
   try {
     const surgery = await Surgery.findById(surgeryId);
 
     if (!surgery) return res.status(404).json({ message: 'Surgery not found' });
 
     await surgery.deleteOne();
+
     res.json({ message: 'Surgery canceled' });
   } catch (err) {
     res.status(500).json({ message: err.message });
